Separate rewind into an abstract class for easier specialization
