+++ Assets/Scripts/Rewinders/IRewinder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dd8156c45c4f04745846c09a5849c748
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
