+++ Assets/Scripts/ZController.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 using UnityEditor.SearchService;
 
 public class ZController : MonoBehaviour {
-    private IEnumerable<Rewind> Rewinds;
+    private IEnumerable<IRewinder> Rewinds;
 
     public float RecordInterval = .2f;
     private float DeltaTime;
@@ -24,17 +24,17 @@ public class ZController : MonoBehaviour {
         if (Active) {
             DeltaTime += Time.deltaTime;
             if (DeltaTime >= (RecordInterval / RewindScale) && (Input.GetMouseButton(0) || Input.GetKey(KeyCode.Q))) {
-                foreach (Rewind r in Rewinds)
+                foreach (IRewinder r in Rewinds)
                     r.RewindState();
                 DeltaTime = 0;
             } else if (DeltaTime >= RecordInterval && Rewinds.Any(r => r.NeedUpdate())) {
-                foreach (Rewind r in Rewinds)
+                foreach (IRewinder r in Rewinds)
                     r.Store();
                 DeltaTime = 0;
             }
 
             if (Input.GetKeyUp(KeyCode.Q) || Input.GetMouseButtonUp(0))
-                foreach (Rewind r in Rewinds)
+                foreach (IRewinder r in Rewinds)
                     r.Play();
         }
     }
@@ -45,14 +45,14 @@ public class ZController : MonoBehaviour {
             z.Active = false;
         }
 
-        foreach (Rewind r in Rewinds)
+        foreach (IRewinder r in Rewinds)
             r.Play();
 
         Active = true;
     }
 
     public void Pause() {
-        foreach (Rewind r in Rewinds)
+        foreach (IRewinder r in Rewinds)
             r.Pause();
     }
 }
